fixing jax test error - using float64

+++ tests/link/jax/test_scalar.py
@@ -169,25 +169,25 @@ def test_tfp_ops(op, test_values):
 
 
 def test_betaincinv():
-    a = vector("a", dtype="float32")
-    b = vector("b", dtype="float32")
-    x = vector("x", dtype="float32")
+    a = vector("a", dtype="float64")
+    b = vector("b", dtype="float64")
+    x = vector("x", dtype="float64")
     out = betaincinv(a, b, x)
     fg = FunctionGraph([a, b, x], [out])
     compare_jax_and_py(fg, [np.array([3.0, 5.5, 0.7])])
 
 
 def test_gammaincinv():
-    k = vector("k", dtype="float32")
-    x = vector("x", dtype="float32")
+    k = vector("k", dtype="float64")
+    x = vector("x", dtype="float64")
     out = gammaincinv(k, x)
     fg = FunctionGraph([k, x], [out])
     compare_jax_and_py(fg, [np.array([5.5, 0.7])])
 
 
 def test_gammainccinv():
-    k = vector("k", dtype="float32")
-    x = vector("x", dtype="float32")
+    k = vector("k", dtype="float64")
+    x = vector("x", dtype="float64")
     out = gammainccinv(k, x)
     fg = FunctionGraph([k, x], [out])
     compare_jax_and_py(fg, [np.array([5.5, 0.7])])
